[MOV] web: move CopyButton to core folder
This commits moves the CopyButton component to the core folders,
as the component was already used outside of fields. This makes the
component generally available as a ClipboardJS replacement option.
