+++ addons/web/static/src/core/copy_button/copy_button.xml
